requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
alembic>=1.18.0
pydantic>=2.0.0
pgvector
rank-bm25>=0.2.2
sentence-transformers>=2.2.0
requests>=2.31.0
cohere>=4.0.0
transformers>=4.30.0
torch>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
google-generativeai>=0.3.0
openai>=1.0.0
jinja2>=3.1.2
aiofiles>=23.2.1
email-validator>=2.0.0
pydantic-settings>=2.0.0
spacy>=3.7.4
thinc>=8.3.2
pydantic>=2.6.4
regex>=2024.5.15
argon2-cffi>=23.1.0
